Merged PR 708: [Bug] Added fix for omnicache and PluginManager
There are two bugs. When no plugins are loaded, a none type is returned for the list of plugins. When we attempt to iterate over the list to init the plugins, we fail since none is not iterable.
The second bug is that when there is no omnicache specified, it defaults to none. When we have a nonetype and we resolve it via the os.path.abspath, it defaults to the working directory. Normally, this is fine since we can use our current repo as a reference. However, on the server in CI environment, the root repo is cloned shallow (10 commits as of time of writing) and git throws an error when this occurs.