Merged PR 873: [Bug Fix] Added safe handler to console logger
Added safe handler to console logger to prevent the logging being initialized early by a logging statement in a utility function. Before, when creating the AnsiHandler for the console log handler, logging was being emitted. This forced the logging library to create it's own handler to deal with the output (which defaults to  a console stream handler). Creating a known safe handler means we can control the creation of a default handler and remove it when we are ready to add the real AnsiHandler. Otherwise, we have two console handlers and one is set at a very low level.