Merged PR 930: Parameterize the UefiBuilder object invoked by CommonBuildEntry.
Previously, CommonBuildEntry blindly imported a module called "PlatformBuildWorker" and failed build if that module wasn't in the path. This is both brittle and inflexible.

This change defaults to looking for PlatformBuildWorker, but allows the caller to pass a module name (including the caller itself) that will be loaded to invoke the next portion of the build process.