Merged PR 926: Add defaults to the CommonBuildEntry for whether to open the log file.
In the CommonBuildEntry "build" process, Python would throw an exception if the UefiBuilder (or subclass) didn't define the LaunchLogOnSuccess or LaunchLogOnError build vars.

To address this, we add a 'default' to VarDict.GetValue(), and pass 'FALSE' as the default for the call in CommonBuildEntry. For all other callers, this should be transparent and GetValue should continue returning None if not found.