Merged PR 1146: Adding UrlDependency
WebDependency takes a source URL to a compressed file online, downloads it, and unzips it (based on filename)
Supports .zip and .tar.* files.
internal_path in ext_dep file is required to explain the internal structure of the compressed file, even if its just the name of the directory that will be extracted.

Taking this opportunity to move ExternalDependency, NugetDependency, and WebDependency to different files. Renaming ExternalDependencies to ExternalDependency.